Login is not persistant
Login is stored with a cookie. When you leave a room or refresh the
page, the server logs you out and then the client logs back in. This
may cause a race condition (especially when clicking the back button)
because we delete the player from the global list on the server then
try to re-add. If the add happens before the delete, the user will get
an error saying his name has already been taken. For now, we should be
safe but this deserves some looking into.